refactor: word cloud 컴포넌트 분리
