+++ src/pages/mainPage/NotificationScreen/NotificationScreen.tsx
@@ -1,12 +1,11 @@
 import React, { useEffect, useRef, useState } from "react";
-import * as echarts from "echarts";
-// import "echarts-wordcloud";
-import carMask from "../../../assets/images/wordcloud.png"; // Make sure this path is correct
+import carMask from "../../../assets/images/wordcloud.png";
 import Button from "../../../components/common/Button/Button";
 import { useNavigate } from "react-router";
 import { EVENT_TERMS } from "../../../constants/EventData";
 import { fetchWordCloudData } from "../../../api/fetch";
 import { WordListResponse } from "../../../types/InfoScreen";
+import WordCloud from "../../../components/mainPage/InfoScreen/WordCloud";
 
 const NotificationScreen = () => {
   const navigate = useNavigate();
@@ -69,45 +68,3 @@ const NotificationScreen = () => {
 };
 
 export default NotificationScreen;
-
-const WordCloud = ({ data, maskImage }: any) => {
-  const chartRef = useRef(null);
-
-  useEffect(() => {
-    const chart = echarts.init(chartRef.current);
-
-    const maskImg = new Image();
-    maskImg.src = maskImage;
-
-    maskImg.onload = function () {
-      const option = {
-        series: [
-          {
-            type: "wordCloud",
-            gridSize: 8,
-            sizeRange: [16, 120],
-            rotationRange: [0, 0],
-            shape: "circle",
-            maskImage: maskImg,
-            textStyle: {
-              fontFamily: "Kia-Signature-Regular",
-              color: function () {
-                const colors = ["#4B7C83", "#4C7DC4", "#C9C9C9"];
-                return colors[Math.floor(Math.random() * colors.length)];
-              },
-            },
-            data: data,
-          },
-        ],
-      };
-
-      chart.setOption(option);
-    };
-
-    return () => {
-      chart.dispose();
-    };
-  }, [data]);
-
-  return <div ref={chartRef} style={{ width: "88rem", height: "36rem" }}></div>;
-};
